requirements.txt
Flask==3.0.3
Flask_Cors==5.0.0
numpy
pandas==2.2.3
#rapidfuzz==3.10.1
Requests==2.32.3
#scikit_learn==1.4.1.post1
#scikit_learn==1.5.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
apscheduler==3.11.0
redis==3.5.3
faiss-cpu==1.10.0
groq==0.15.0
llama_index==0.12.13
llama_index.llms.groq==0.3.1
#langchain_groq==0.2.4
#langchain_core==0.3.39
sentence_transformers==3.4.0
gunicorn
llama-index-embeddings-huggingface==0.5.4
onnxruntime==1.22.0
langchain-groq==0.3.2
python-docx==1.1.2
langchain_community==0.3.23
requests==2.32.3
gdown==5.2.0
pymupdf==1.25.5
openpyxl==3.1.5
# must install https://aka.ms/vs/17/release/vc_redist.x64.exe